[dram_perf, config] Specify the 'dram_access_cost' and 'offchip_bandwidth' in 'ns' and 'GB/s' respectively

+++ common/performance_model/dram_perf_model.h
@@ -23,7 +23,7 @@ class DramPerfModel
       QueueModel* m_queue_model;
       Core* m_core;
       MovingAverage<UInt64>* m_moving_average;
-      UInt32 m_dram_access_cost;
+      float m_dram_access_cost;
       float m_dram_bandwidth;
 
       bool m_queueing_model_enabled;
